backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
apscheduler==3.10.4
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.6.4
loguru==0.7.2
PyJWT==2.8.0
cryptography==42.0.0
slowapi==0.1.9

# Test dependencies
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
